Tests/LibWeb/Text/expected/abspos-relayout-move-margin-change.txt

boundary=35,25,100,50
partial-layouts=0 full-layouts=1
FAIL
